Codebase quality audit: fix swallowed errors, bugs, and structural issues (closes #26)
